requirements-inference.txt
accelerate==0.24.1
diffusers==0.25.1
einops==0.8.1
huggingface_hub==0.23.5
imageio==2.34.2
jaxtyping==0.2.38
numpy==2.2.3
omegaconf==2.3.0
opencv_python==4.10.0.84
packaging==24.2
peft==0.7.0
Pillow==11.1.0
pytorch_lightning==2.5.0.post0
safetensors==0.5.3
torch==1.13.1
torchvision==0.14.1
tqdm==4.66.1
transformers==4.40.1
typeguard==4.4.1
xformers==0.0.22.post7
